base works

+++ src/screens/AdminDashboard/QuestionnaireManagement.jsx
@@ -1,4 +1,3 @@
-// src/components/QuestionnaireManagement/QuestionnaireManagement.jsx
 import React, { useEffect, useState } from 'react'
 import { db } from '../../firebase/config'
 import {
@@ -19,8 +18,19 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faFloppyDisk, faBars } from '@fortawesome/free-solid-svg-icons'
 import LoadingSpinner from '../../components/LoadingSpinner/LoadingSpinner'
 import EditScoreDescModal from '../../components/helpScore/EditScoreDescModal'
+import { Tooltip, TextField } from '@mui/material'
 
-Modal.setAppElement('#root') // Adjust this selector to your app's root element
+Modal.setAppElement('#root')
+
+const ScoreTooltip = ({ children, error }) => (
+	
+	<Tooltip title="ניקוד זה כבר קיים. אנא בחרי ניקוד אחר." open={error}
+	classes={{ tooltip: 'custom-tooltip' }}
+>
+		
+		{children}
+	</Tooltip>
+)
 
 const QuestionnaireManagement = ({ questionnaireId }) => {
 	const [description, setDescription] = useState('')
@@ -41,7 +51,6 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 	const [editingAnswer, setEditingAnswer] = useState(null) // New state for editing answer
 	const [showNewAnswerFields, setShowNewAnswerFields] = useState(false)
 	const [editScoreDescModalIsOpen, setEditScoreDescModalIsOpen] = useState(false);
-	
 
 	// for loading icon:
 	const [isReordering, setIsReordering] = useState(false)
@@ -165,13 +174,16 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 		setModalIsOpen(true)
 	}
 
+	const checkDuplicateScore = (score, answers) => {
+		return answers.some((answer) => answer.score === parseInt(score, 10))
+	}
+
 	const handleAddAnswer = async () => {
 		if (newAnswerText && newAnswerScore) {
 			const score = parseInt(newAnswerScore, 10)
-			const scoreExists = selectedQuestion.answers.some((answer) => answer.score === score)
+			const scoreExists = checkDuplicateScore(score, selectedQuestion.answers)
 
 			if (scoreExists) {
-				alert('This score already exists. Please choose a different score.')
 				return
 			}
 
@@ -187,77 +199,78 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 			setSelectedQuestion(updatedQuestion)
 
 			// Update Firebase
-			try {
-				const questionDocRef = doc(db, 'Questionnaire', selectedQuestion.id)
-				await updateDoc(questionDocRef, {
-					[newAnswerScore]: newAnswerText,
-				})
+			// check if we need to update(if any changes were made)
+			if (JSON.stringify(selectedQuestion.answers) !== JSON.stringify(updatedAnswers)) {
 
-				// Reset the input fields
-				setNewAnswerText('')
-				setNewAnswerScore('')
-				alert('Answer added successfully!')
-			} catch (error) {
-				handleFirestoreError(error)
+				try {
+					const questionDocRef = doc(db, 'Questionnaire', selectedQuestion.id)
+					await updateDoc(questionDocRef, {
+						[newAnswerScore]: newAnswerText,
+					})
+
+					// Reset the input fields
+					setNewAnswerText('')
+					setNewAnswerScore('')
+				} catch (error) {
+					handleFirestoreError(error)
+				}
 			}
+
 		}
 	}
-
 
 	const handleEditAnswer = (answer) => {
 		setEditingAnswer(answer)
 		setCurrentText(answer.text)
 		setCurrentScore(answer.score)
 	}
-	
+
 	const saveEditedAnswer = async () => {
-		const previousScore = editingAnswer.score;
+		// const previousScore = editingAnswer.score;
 		const updatedAnswers = selectedQuestion.answers.map((answer) =>
 			answer.id === editingAnswer.id
 				? { ...answer, text: currentText, score: parseInt(currentScore, 10) }
 				: answer
 		);
-	
+
+		// Check for duplicate scores
+		const scoreExists = checkDuplicateScore(currentScore, updatedAnswers.filter(a => a.id !== editingAnswer.id))
+		if (scoreExists) {
+			// don't save the answer, show an error message
+			return
+			
+		}
+
 		// Sort the updatedAnswers array by score
 		updatedAnswers.sort((a, b) => a.score - b.score);
-	
+
 		const updatedQuestion = { ...selectedQuestion, answers: updatedAnswers };
 		setSelectedQuestion(updatedQuestion);
 		setEditingAnswer(null);
-	
+
 		// Update Firebase
-		try {
-			const questionDocRef = doc(db, 'Questionnaire', selectedQuestion.id);
-			const questionDoc = await getDoc(questionDocRef);
-			const questionData = questionDoc.data();
-	
-			// Check if the new score already exists
-			const scoreExists = Object.keys(questionData).some(
-				(key) => key !== editingAnswer.id && parseInt(key, 10) === parseInt(currentScore, 10)
-			);
-	
-			if (scoreExists) {
-				// Here clown
-				alert('This score already exists. Please choose a different score.');
-				return;
+		// check if we need to update(if any changes were made)
+		if (JSON.stringify(selectedQuestion.answers) !== JSON.stringify(updatedAnswers)) {
+			try {
+				const questionDocRef = doc(db, 'Questionnaire', selectedQuestion.id);
+				const questionDoc = await getDoc(questionDocRef);
+				const questionData = questionDoc.data();
+
+				// Remove the old answer
+				await updateDoc(questionDocRef, {
+					[editingAnswer.id]: deleteField(),
+				});
+
+				// Add the updated answer with the new score as the key
+				await updateDoc(questionDocRef, {
+					[currentScore]: currentText,
+				});
+
+			} catch (error) {
+				handleFirestoreError(error);
 			}
-	
-			// Remove the old answer
-			await updateDoc(questionDocRef, {
-				[editingAnswer.id]: deleteField(),
-			});
-	
-			// Add the updated answer with the new score as the key
-			await updateDoc(questionDocRef, {
-				[currentScore]: currentText,
-			});
-	
-			alert('Answer updated successfully!');
-		} catch (error) {
-			handleFirestoreError(error);
 		}
 	};
-	
 
 	const saveQuestionChanges = async () => {
 		const { id, question, required, answers } = selectedQuestion
@@ -274,7 +287,7 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 			)
 			setModalIsOpen(false)
 			resetInputFields()
-			alert('Question updated successfully!')
+			alert('השאלה עודכנה בהצלחה!')
 		} catch (error) {
 			handleFirestoreError(error)
 		}
@@ -299,7 +312,6 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 	}
 
 	const handleAnswerChange = (e) => {
-		// clown
 		const { name, value } = e.target
 		if (name === 'answerText') {
 			setCurrentText(value)
@@ -458,7 +470,7 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 								הוספת שאלה
 							</button>
 							<button className='edit-score-desc-button'
-							onClick={() => setEditScoreDescModalIsOpen(true)}>
+								onClick={() => setEditScoreDescModalIsOpen(true)}>
 								עריכת תיאורי תוצאות המבדק
 							</button>
 
@@ -586,12 +598,15 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 															</td>
 															<td>
 																{editingAnswer && editingAnswer.id === answer.id ? (
-																	<input
-																		type='number'
-																		name='answerScore'
-																		value={currentScore}
-																		onChange={handleAnswerChange}
-																	/>
+																	<ScoreTooltip error={checkDuplicateScore(currentScore, selectedQuestion.answers.filter(a => a.id !== editingAnswer.id))}>
+																		<TextField
+																			type='number'
+																			name='answerScore'
+																			value={currentScore}
+																			onChange={handleAnswerChange}
+																			error={checkDuplicateScore(currentScore, selectedQuestion.answers.filter(a => a.id !== editingAnswer.id))}
+																		/>
+																	</ScoreTooltip>
 																) : (
 																	<>ניקוד: {answer.score}</>
 																)}
@@ -639,12 +654,15 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 															/>
 														</td>
 														<td>
-															<input
-																type='number'
-																placeholder='ניקוד'
-																value={newAnswerScore}
-																onChange={(e) => setNewAnswerScore(e.target.value)}
-															/>
+															<ScoreTooltip error={checkDuplicateScore(newAnswerScore, selectedQuestion.answers)}>
+																<TextField
+																	type='number'
+																	placeholder='ניקוד'
+																	value={newAnswerScore}
+																	onChange={(e) => setNewAnswerScore(e.target.value)}
+																	error={checkDuplicateScore(newAnswerScore, selectedQuestion.answers)}
+																/>
+															</ScoreTooltip>
 														</td>
 														<td>
 															<button
@@ -707,7 +725,7 @@ const QuestionnaireManagement = ({ questionnaireId }) => {
 					<EditScoreDescModal
 						isOpen={editScoreDescModalIsOpen}
 						onRequestClose={() => setEditScoreDescModalIsOpen(false)}
-						
+
 					/>
 
 				</>
